+++ CHANGELOG.md
@@ -5,6 +5,7 @@ mdr1986x-small-lib changelog
 ### Added
 - Added HSE and clock source type parameters in SYS_ClkInit for flexible clock
 setup. Changed version of system.c to 0.4.
+- Added some frequency values in SYS_Freq_Enum
 
 ## [0.1] - 2019-06-25
 ### Added